Update README.md
This is my Bike sales project on Excel. In this Excel  file I created a welcome page, Index page, and few pages with chart and pivot tables and then last dashboard for analyzing data. The main objective of this Bike Sales project is to make a dashboard to visualize and analyze with different field. This project excel sheet consist of 8 pages. First page is welcome page of project and a link is given to it which help to go to the next page. The second page is Index which includes all the links for opening our desired pages. bike_buyers is the page were the date in imported and the data that is cleaned is given in the page DATA. Now pivot table and its pivot chart pie chart, column chart and line chart. All the charts are given to a single page dashboard and slicers are also applied in this page for proper analyzing. If we click on slicer then the charts that created automatically changes .
The first chart which consist of bike purchased people by gender with a column chart. From this chart we can analyze that the bike purchased by males is more than that of females. The second chart is about the age wise bike purchased data. Pie chart is used here for visualize through the pie chart itself it clear that the middle aged people are purchased more than youth and senior. Next chart is line chart to visualize the data count of purchased. In this chart we can analyze that customers tend to purchase bike more for short commutes (0-1 miles)than longer commuters where more then 10 miles. These are key insights from the Bike Sales Dashboard